backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
pydantic
sqlalchemy
psycopg2-binary
python-dotenv
